interpretation: split graph into read/write submodules
Per Nico's #883 review: graph.rs had grown past 1000 lines (with its
co-located tests). Break it into a `graph/` module along the read/write seam,
keeping each function's tests in the same file:

- graph/mod.rs   — shared core: ExistingInstances + InstanceContext, the
  class/identity/relation helpers (instances_by_class, identity_values_by_class,
  identity_property, normalize_identity, relation_names/predicates,
  class_local_name), re-exporting read + write.
- graph/read.rs  — pulling state from the perspective: existing_instance_context,
  gather_transcript(_sparql), decode_literal_string (+ their tests).
- graph/write.rs — planning writes: InterpretationOp, plan_interpretation_ops
  (_with_context), relation resolution, scalar_values (+ their tests).

Also moved `filter_already_present` (the NormalizedString dedup arm) out of
graph and into dedup.rs alongside the other two dedup entry points and its
tests — it's dedup, not graph, and this puts the function back with its tests.

Pure move, no logic changes. graph.rs (1387 lines) → mod 132 / read 401 /
write 789. All 45 interpretation tests pass (--test-threads=1); no new warnings.
