+++ rust-executor/src/perspectives/interpretation/graph/write.rs
@@ -1,187 +1,9 @@
-use super::ProposedInstance;
-use crate::perspectives::model_query::types::{ModelShape, ParentScope, ShapeProperty};
-use crate::perspectives::perspective_instance::PerspectiveInstance;
+use super::*;
+use crate::perspectives::interpretation::ProposedInstance;
+use crate::perspectives::model_query::types::ModelShape;
 use crate::types::Link;
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use uuid::Uuid;
-
-/// The existing instances in scope for an interpretation pass, keyed by base
-/// URI (`id`). This is the **single source of truth** the whole pass reads:
-/// the prompt view ([`build_interpretation_input`]), the deterministic dedup
-/// safety net ([`filter_already_present`] / semantic), and Create-vs-Update
-/// routing ([`plan_interpretation_ops_with_context`]) all project what they
-/// need from this one map — instead of separately-threaded `class → identity`
-/// and `id-set` views that could drift out of sync. Each [`InstanceContext`]
-/// still carries its own `id`, so the value is self-describing; the key is
-/// that same id, promoted for O(1) "does the graph hold this id?" checks.
-pub type ExistingInstances = HashMap<String, InstanceContext>;
-
-/// Group existing instances by class local name for the per-class reasoning
-/// the prompt and dedup paths do. Each class's instances are sorted by `id` so
-/// the prompt text and the semantic-embedding batch order are deterministic
-/// across runs (an id-keyed map has no inherent order).
-pub(crate) fn instances_by_class(
-    existing: &ExistingInstances,
-) -> BTreeMap<String, Vec<&InstanceContext>> {
-    let mut out: BTreeMap<String, Vec<&InstanceContext>> = BTreeMap::new();
-    for inst in existing.values() {
-        out.entry(inst.class.clone()).or_default().push(inst);
-    }
-    for rows in out.values_mut() {
-        rows.sort_by(|a, b| a.id.cmp(&b.id));
-    }
-    out
-}
-
-/// Per-class raw identity values (titles) projected from the id-keyed existing
-/// set — the comparison basis for both dedup paths (the string path normalizes
-/// them, the semantic path embeds them). Deterministic order (by `id`) so the
-/// semantic embedding batch is stable.
-pub(crate) fn identity_values_by_class(
-    existing: &ExistingInstances,
-) -> HashMap<String, Vec<String>> {
-    instances_by_class(existing)
-        .into_iter()
-        .map(|(class, rows)| (class, rows.into_iter().map(|r| r.title.clone()).collect()))
-        .collect()
-}
-
-/// The property a class declares as its dedup identity (its title-like
-/// interpretation key) — the first property with `identity == true`. `None`
-/// when the SDNA declared no identity, in which case the class is never
-/// deduplicated (still interpreted and created, just not deduped).
-pub(crate) fn identity_property(shape: &ModelShape) -> Option<&ShapeProperty> {
-    shape.properties.iter().find(|p| p.identity)
-}
-
-/// Canonicalize an identity value for equality: trim, collapse internal
-/// whitespace to single spaces, and lowercase. So "Ship  the MVP " and
-/// "ship the mvp" compare equal. Semantic/embedding dedup is a later
-/// follow-up; this is deliberately a cheap normalized string match.
-pub(crate) fn normalize_identity(s: &str) -> String {
-    s.split_whitespace()
-        .collect::<Vec<_>>()
-        .join(" ")
-        .to_lowercase()
-}
-
-/// Names of the shape's relation (link-typed) properties. `load_shape` lists
-/// every relation both in `properties` (so the query pipeline sees its
-/// predicate) *and* in `include_relations`; we key off the latter to recognise
-/// them. Relations never travel through the scalar write path: their targets are
-/// instance URIs, not literals, so `create_subject`/`update_subject` must not see
-/// them in `initial_values` (a setter would literal-encode an instance ref into a
-/// bogus `literal:` URI). They are resolved separately into
-/// [`InterpretationOp::AddLinks`].
-pub(crate) fn relation_names(shape: &ModelShape) -> HashSet<&str> {
-    shape
-        .include_relations
-        .iter()
-        .map(|r| r.name.as_str())
-        .collect()
-}
-
-/// Predicates of the shape's relation properties — the `properties`-side view of
-/// [`relation_names`], used where a shape property must be recognised as a
-/// relation by predicate (prompt field rendering, hint join).
-pub(crate) fn relation_predicates(shape: &ModelShape) -> HashSet<&str> {
-    shape
-        .include_relations
-        .iter()
-        .map(|r| r.predicate.as_str())
-        .collect()
-}
-
-/// Deterministic dedup safety-net (pure): drop proposed instances whose
-/// (class, identity-value) already exists in the graph, compared under
-/// [`normalize_identity`]. This is the hard guarantee behind the soft
-/// `existing` hint in [`build_interpretation_input`] — even if the model
-/// re-proposes a known item, it never becomes a new instance.
-///
-/// This is the **`DedupStrategy::NormalizedString`** implementation, not a
-/// separate/legacy path: `run_interpretation` always dedups through
-/// [`filter_already_present_with_strategy`](super::filter_already_present_with_strategy),
-/// which calls this for the (default) string strategy and the embedding path
-/// for `DedupStrategy::Semantic`. No production caller invokes this directly.
-///
-/// `existing` is the id-keyed [`ExistingInstances`] source of truth; the
-/// per-class identity values are projected from it here. `identity_props` maps
-/// a class's local name to the NAME of its declared identity property. An
-/// instance whose class has no identity property is always kept (no dedup);
-/// likewise one missing that property's value.
-///
-/// Filters **in place**: the surviving instances keep the LLM's emission order,
-/// which is what the `new:<Class>:<n>` relation ordinals in
-/// [`plan_interpretation_ops_with_context`] resolve against.
-pub fn filter_already_present(
-    instances: Vec<ProposedInstance>,
-    existing: &ExistingInstances,
-    identity_props: &HashMap<String, String>,
-) -> Vec<ProposedInstance> {
-    // Seed per-class known sets with pre-existing identities; each accepted
-    // proposal is added to its class's set so a same-response duplicate is
-    // dropped like an already-persisted one. Without this, an LLM that emits
-    // the same (class, identity) twice would slip through and `run_interpretation`
-    // would create two subjects for it.
-    let mut known: HashMap<String, HashSet<String>> = identity_values_by_class(existing)
-        .into_iter()
-        .map(|(class, values)| {
-            (
-                class,
-                values.iter().map(|v| normalize_identity(v)).collect(),
-            )
-        })
-        .collect();
-    let mut out = Vec::with_capacity(instances.len());
-    for inst in instances {
-        // A *trusted* id is an explicit upsert target — it names a specific
-        // existing node, so its identity value *should* match one already
-        // present; never dedup it away. But the planner only routes to Update
-        // for ids the graph actually holds; a hallucinated id absent from
-        // `existing` would be routed to Create, so it must still be
-        // dedup-checked like an ordinary proposal (else a made-up id +
-        // duplicate identity mints a duplicate node).
-        if inst
-            .id
-            .as_deref()
-            .is_some_and(|id| existing.contains_key(id))
-        {
-            out.push(inst);
-            continue;
-        }
-        // No declared identity for this class ⇒ no dedup.
-        let Some(idp_name) = identity_props.get(&inst.class) else {
-            out.push(inst);
-            continue;
-        };
-        let Some(value) = inst.props.get(idp_name).and_then(|v| v.as_str()) else {
-            // no identity value to compare on — keep it
-            out.push(inst);
-            continue;
-        };
-        let normalized = normalize_identity(value);
-        let set = known.entry(inst.class.clone()).or_default();
-        if set.contains(&normalized) {
-            log::debug!(
-                "interpretation: dropping already-present {} '{}'",
-                inst.class,
-                value
-            );
-            continue;
-        }
-        set.insert(normalized);
-        out.push(inst);
-    }
-    out
-}
-
-/// Local class name from a class URI: `ns://Intention` -> `Intention`.
-pub(crate) fn class_local_name(target_class: &str) -> &str {
-    target_class
-        .rsplit(|c| c == '/' || c == ':')
-        .find(|seg| !seg.is_empty())
-        .unwrap_or(target_class)
-}
 
 /// The scalar (non-relation) field values a proposed instance wants written —
 /// the payload handed to `create_subject` / `update_subject` as `initial_values`.
@@ -478,263 +300,6 @@ fn resolve_ref(
     } else {
         None
     }
-}
-
-/// minimal transcript gatherer. Reads links `source ⇒ predicate ⇒ literal`
-/// from a perspective where `predicate` matches `message_predicate` and the
-/// target is a `literal:string:` URI (i.e., a message body). Returns turns in
-/// the order the store returned them. Speaker is the link author.
-///
-/// Kept intentionally small — flows/channel-aware traversal is deferred to a
-/// later PR. Callers that already have a curated `Vec<(speaker, text)>` should
-/// pass it straight to [`run_interpretation`] and skip this helper.
-pub async fn gather_transcript(
-    perspective: &PerspectiveInstance,
-    source: &str,
-    message_predicate: &str,
-) -> anyhow::Result<Vec<(String, String)>> {
-    use crate::types::LinkQuery;
-    let query = LinkQuery {
-        source: Some(source.to_string()),
-        predicate: Some(message_predicate.to_string()),
-        ..Default::default()
-    };
-    let links = perspective
-        .get_links(&query)
-        .await
-        .map_err(|e| anyhow::anyhow!("gather_transcript: get_links failed: {e:#}"))?;
-    let mut out = Vec::with_capacity(links.len());
-    for l in links {
-        if let Some(body) = decode_literal_string(&l.data.target) {
-            out.push((l.author, body));
-        }
-    }
-    Ok(out)
-}
-
-/// Decode a `literal:` URI to a plain `String`, but only when it holds a
-/// string value. Reuses `ad4m-client`'s `Literal` (the canonical
-/// parse/serialize type); non-string literals (number/bool/json) and
-/// non-literal URIs yield `None`.
-pub(crate) fn decode_literal_string(uri: &str) -> Option<String> {
-    use ad4m_client::literal::{Literal, LiteralValue};
-    match Literal::from_url(uri.to_string()).ok()?.get().ok()? {
-        LiteralValue::String(s) => Some(s),
-        _ => None,
-    }
-}
-
-/// Assemble an interpretation transcript from an arbitrary SPARQL SELECT
-/// against the perspective. The query MUST bind `?speaker` and `?text` in
-/// its solutions — each row becomes a `(speaker, text)` turn in the order
-/// the store returned them. `?text` may be either a plain string literal
-/// (returned as-is) or a `literal:string:...` URI (decoded via
-/// [`decode_literal_string`]); rows that decode to non-string literals or
-/// bind neither cleanly are skipped rather than failing the whole gather.
-///
-/// This is the generic counterpart to [`gather_transcript`]: a caller supplies
-/// an arbitrary scope query (the AutoProcessor reads one off its config; tests
-/// pass one directly) so interpretation can run over just the relevant subset
-/// of the perspective instead of the full channel. Callers that just want the
-/// flat "all messages on this source under this predicate" view should keep
-/// using the source+predicate wrapper.
-pub async fn gather_transcript_sparql(
-    perspective: &PerspectiveInstance,
-    sparql: &str,
-) -> anyhow::Result<Vec<(String, String)>> {
-    let rows_json = perspective
-        .sparql_query(sparql.to_string())
-        .map_err(|e| anyhow::anyhow!("gather_transcript_sparql: SPARQL query failed: {e:#}"))?;
-    let rows: Vec<serde_json::Value> = serde_json::from_str(&rows_json)
-        .map_err(|e| anyhow::anyhow!("gather_transcript_sparql: bad SPARQL result JSON: {e:#}"))?;
-
-    let mut out = Vec::with_capacity(rows.len());
-    for row in rows {
-        let Some(speaker) = row.get("speaker").and_then(|v| v.as_str()) else {
-            continue;
-        };
-        let Some(raw_text) = row.get("text").and_then(|v| v.as_str()) else {
-            continue;
-        };
-        let text = if raw_text.starts_with("literal:") {
-            match decode_literal_string(raw_text) {
-                Some(s) => s,
-                None => continue, // non-string literal — not a message body
-            }
-        } else {
-            raw_text.to_string()
-        };
-        out.push((speaker.to_string(), text));
-    }
-    Ok(out)
-}
-
-/// One existing instance the interpreter should know about — the LLM sees these
-/// so it can decide whether an interpreted item is a genuinely new node (no `id`
-/// on the output) or the continuation/refinement of an existing one (emit this
-/// entry's `id` to trigger the upsert path in
-/// [`plan_interpretation_ops_with_context`]).
-///
-/// `class` is redundant with the enclosing map key, but kept on each row so the
-/// JSON entry rendered into the prompt is self-contained and unambiguous when
-/// the LLM scans a mixed-class list.
-#[derive(Debug, Clone, PartialEq)]
-pub struct InstanceContext {
-    /// Base URI of the existing instance — what the LLM emits as `id` to update.
-    pub id: String,
-    /// The class's declared `identity` value (usually `title`), decoded. Raw,
-    /// not normalized: the prompt shows it to the LLM verbatim, and
-    /// [`filter_already_present`] normalizes both sides when comparing.
-    pub title: String,
-    /// Local class name (e.g. "Task"), matching the map key of the returned map.
-    pub class: String,
-    /// Currently-set secondary scalar values, keyed by property name. Excludes
-    /// the identity property (already rendered as `title`), the class's type
-    /// flag, and every relation. Empty when the class declares no other
-    /// scalars, or when this instance has none set. Rendered into the prompt
-    /// so the LLM sees the existing instance's *state* — not just its
-    /// identity label — and can better judge whether a new turn continues an
-    /// existing instance or belongs to a fresh one on a different topic.
-    /// `BTreeMap` for deterministic prompt ordering across calls.
-    pub properties: BTreeMap<String, String>,
-}
-
-/// read the instances already present in the perspective for each target class
-/// into the id-keyed [`ExistingInstances`] map. Each [`InstanceContext`] carries
-/// the instance's `id` (base URI) alongside its declared identity value and
-/// secondary scalars.
-///
-/// This one map serves both halves of the interpretation contract; consumers
-/// project what they need from it (see [`instances_by_class`] /
-/// [`identity_values_by_class`]):
-///   * the prompt ([`build_interpretation_input`]) shows `id` + identity so the
-///     model can steer away from re-proposing known items *and* emit an `id` to
-///     upsert one, or reference it as a relation target;
-///   * the identity values feed the deterministic dedup safety net
-///     ([`filter_already_present`]), and the id key-set bounds which existing-id
-///     relation refs are accepted ([`plan_interpretation_ops_with_context`]).
-///
-/// The dedup/display key is whichever property the class declares as its
-/// `identity` (via [`identity_property`]), not a hard-coded `title`. Classes
-/// with no identity property are skipped entirely — no identity ⇒ no dedup, and
-/// nothing meaningful to show the model.
-///
-/// Instances are read through the model-query API (`PerspectiveInstance::
-/// model_query`) — the symmetric counterpart to writing them via
-/// `create_subject` — so class conformance and field decoding go through the
-/// class's own shape/getters rather than hand-matched type-flag links.
-///
-/// A per-class `model_query` failure is propagated. Silently treating it as
-/// "no existing instances" would break [`filter_already_present`]'s deterministic
-/// dedup guarantee: an empty `known` set for the failing class lets the LLM's
-/// re-proposal of an existing item slip through and mint a duplicate subject.
-/// Callers can exclude affected classes upstream (via `interpretation_classes`)
-/// if they want a soft-skip.
-pub async fn existing_instance_context(
-    perspective: &PerspectiveInstance,
-    shapes: &[ModelShape],
-    scope: Option<&ParentScope>,
-) -> anyhow::Result<ExistingInstances> {
-    let mut out: ExistingInstances = HashMap::new();
-    for shape in shapes {
-        // No declared identity property ⇒ no dedup key ⇒ skip.
-        let Some(idp) = identity_property(shape) else {
-            continue;
-        };
-        let idp_name = idp.name.clone();
-        let class = class_local_name(&shape.target_class);
-
-        // Secondary scalars: everything the LLM can meaningfully see about an
-        // existing instance beyond its identity — excluding the class type
-        // flag (setter-managed) and every relation (link-typed, hydrated
-        // separately and not useful in the compact `existing` prompt view).
-        // Kept as `Vec<String>` (not a set) so property order is stable across
-        // calls, feeding a deterministic prompt.
-        let rel_preds = relation_predicates(shape);
-        let scalar_names: Vec<String> = shape
-            .properties
-            .iter()
-            .filter(|p| {
-                p.name != idp_name && !p.is_flag && !rel_preds.contains(p.predicate.as_str())
-            })
-            .map(|p| p.name.clone())
-            .collect();
-
-        let mut requested = Vec::with_capacity(1 + scalar_names.len());
-        requested.push(idp_name.clone());
-        requested.extend(scalar_names.iter().cloned());
-        // Build the model-query as structured JSON so an optional `scope`
-        // (a `ParentScope`) can be spliced in as the query's `parent` filter —
-        // constraining the existing-instance set to a sub-graph/tree rather
-        // than every instance of the class in the perspective. `None` keeps the
-        // pre-scope behaviour (all instances). The actual scope is supplied at
-        // runtime by the caller (the AutoProcessor wires it in #885); here we
-        // only thread the plumbing and exercise it in tests.
-        let mut query_obj = serde_json::json!({ "properties": requested });
-        if let Some(scope) = scope {
-            query_obj["parent"] = serde_json::to_value(scope).map_err(|e| {
-                anyhow::anyhow!("existing_instance_context: serialize parent scope: {e:#}")
-            })?;
-        }
-        let query = query_obj.to_string();
-        let result_json = perspective.model_query(class, &query).await.map_err(|e| {
-            anyhow::anyhow!(
-                "existing_instance_context: model_query({class}) failed — refusing to \
-                 proceed because an empty existing-set here would silently break dedup: {e:#}"
-            )
-        })?;
-        let result: serde_json::Value = serde_json::from_str(&result_json).map_err(|e| {
-            anyhow::anyhow!("existing_instance_context: bad model_query result for {class}: {e:#}")
-        })?;
-
-        let rows: Vec<InstanceContext> = result
-            .get("instances")
-            .and_then(|v| v.as_array())
-            .map(|rows| {
-                rows.iter()
-                    .filter_map(|inst| {
-                        // `id` is the instance's base URI, injected by the
-                        // model-query hydration for every row.
-                        let id = inst.get("id").and_then(|v| v.as_str())?;
-                        let title = inst.get(&idp_name).and_then(|v| v.as_str())?;
-                        let mut properties: BTreeMap<String, String> = BTreeMap::new();
-                        for name in &scalar_names {
-                            // Present only when hydrated to a non-empty
-                            // string. Numeric/bool scalars are stringified so
-                            // the prompt stays JSON-object-of-strings for the
-                            // LLM (matches how the model already renders
-                            // interpreted field values on the way in).
-                            let v = inst.get(name);
-                            let rendered = match v {
-                                Some(serde_json::Value::String(s)) if !s.is_empty() => {
-                                    Some(s.clone())
-                                }
-                                Some(serde_json::Value::Number(n)) => Some(n.to_string()),
-                                Some(serde_json::Value::Bool(b)) => Some(b.to_string()),
-                                _ => None,
-                            };
-                            if let Some(rendered) = rendered {
-                                properties.insert(name.clone(), rendered);
-                            }
-                        }
-                        Some(InstanceContext {
-                            id: id.to_string(),
-                            title: title.to_string(),
-                            class: class.to_string(),
-                            properties,
-                        })
-                    })
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        // Key each instance by its own base URI (the single-source id-keyed
-        // map). Ids are globally unique across classes, so no collision.
-        for row in rows {
-            out.insert(row.id.clone(), row);
-        }
-    }
-    Ok(out)
 }
 
 #[cfg(test)]
@@ -1220,168 +785,5 @@ mod tests {
             ],
             "AddLinks must accumulate, not replace"
         );
-    }
-
-    #[tokio::test]
-    async fn existing_instance_context_reads_id_and_identity() {
-        // The context snapshot the prompt + relation resolver rely on: one row per
-        // persisted instance, carrying the base URI as `id` and the class's declared
-        // identity value as `title`.
-        let (mut perspective, shapes, ctx) = setup_perspective_no_llm(&[("Task", TASK_SDNA)]).await;
-        seed_instance(
-            &mut perspective,
-            &ctx,
-            &shapes[0],
-            "soa://existing/task/1",
-            "Migrate the SHACL parser",
-        )
-        .await;
-
-        let ctx_map = existing_instance_context(&perspective, &shapes, None)
-            .await
-            .expect("existing_instance_context");
-        assert_eq!(ctx_map.len(), 1, "one seeded instance; got {ctx_map:#?}");
-        // Id-keyed single source: look the instance up by its base URI.
-        let inst = ctx_map
-            .get("soa://existing/task/1")
-            .expect("instance keyed by its id");
-        assert_eq!(inst.id, "soa://existing/task/1");
-        assert_eq!(inst.title, "Migrate the SHACL parser");
-        assert_eq!(inst.class, "Task");
-        // Task's `owner` scalar is unset on this seed → nothing to render;
-        // stays empty so the prompt does not carry an empty `properties` block.
-        assert!(
-            inst.properties.is_empty(),
-            "task with only identity set must carry no secondary scalars; got {:?}",
-            inst.properties
-        );
-
-        // The projections consumers derive from this one map: per-class identity
-        // values (dedup net) and the id key-set (relation resolver).
-        assert_eq!(
-            identity_values_by_class(&ctx_map).get("Task"),
-            Some(&vec!["Migrate the SHACL parser".to_string()])
-        );
-        assert!(ctx_map.contains_key("soa://existing/task/1"));
-    }
-
-    #[tokio::test]
-    async fn existing_instance_context_scope_constrains_to_subtree() {
-        // Plumbing for tree-scoped extraction (#883): the existing-instance snapshot
-        // that feeds dedup must be constrainable to a sub-graph, not every instance
-        // of the class in the perspective. The runner (#885) supplies the scope; here
-        // we prove the parameter actually narrows the set. Two Tasks, each linked
-        // under a different parent node.
-        use crate::perspectives::model_query::types::ParentScope;
-        use crate::types::{Link, LinkStatus};
-
-        let (mut perspective, shapes, ctx) = setup_perspective_no_llm(&[("Task", TASK_SDNA)]).await;
-        seed_instance(
-            &mut perspective,
-            &ctx,
-            &shapes[0],
-            "soa://tree-a/task/1",
-            "Task under tree A",
-        )
-        .await;
-        seed_instance(
-            &mut perspective,
-            &ctx,
-            &shapes[0],
-            "soa://tree-b/task/1",
-            "Task under tree B",
-        )
-        .await;
-        // Parent edges: <parent> ns://contains <task-base>.
-        for (parent, task) in [
-            ("soa://parent/a", "soa://tree-a/task/1"),
-            ("soa://parent/b", "soa://tree-b/task/1"),
-        ] {
-            perspective
-                .add_link(
-                    Link {
-                        source: parent.into(),
-                        predicate: Some("ns://contains".into()),
-                        target: task.into(),
-                    },
-                    LinkStatus::Local,
-                    None,
-                    &ctx,
-                )
-                .await
-                .expect("parent link");
-        }
-
-        // Unscoped: both instances are visible (pre-scope behaviour).
-        let all = existing_instance_context(&perspective, &shapes, None)
-            .await
-            .expect("unscoped context");
-        assert_eq!(
-            all.len(),
-            2,
-            "unscoped context must see both tasks; got {all:#?}"
-        );
-
-        // Scoped to parent A: only the task under tree A survives.
-        let scope = ParentScope::Raw {
-            id: "soa://parent/a".into(),
-            predicate: "ns://contains".into(),
-        };
-        let scoped = existing_instance_context(&perspective, &shapes, Some(&scope))
-            .await
-            .expect("scoped context");
-        assert_eq!(
-            scoped.len(),
-            1,
-            "parent-scoped context must exclude tree B; got {scoped:#?}"
-        );
-        let inst = scoped
-            .get("soa://tree-a/task/1")
-            .expect("tree-A task present");
-        assert_eq!(inst.id, "soa://tree-a/task/1");
-        assert_eq!(inst.title, "Task under tree A");
-    }
-
-    #[tokio::test]
-    async fn existing_instance_context_populates_secondary_scalars() {
-        // With the design-fork (i) enrichment, `existing_instance_context` must
-        // return each instance's currently-set non-identity scalar values in the
-        // `properties` map, so the prompt can show the LLM the full state of an
-        // existing entry (its rolling summary etc.) — not just the identity
-        // label. This is what unblocks the topic-shift discrimination on smaller
-        // local models.
-        let (mut perspective, shapes, ctx) =
-            setup_perspective_no_llm(&[("ConversationSubgroup", CONVERSATION_SUBGROUP_SDNA)]).await;
-        seed_instance_with_props(
-        &mut perspective,
-        &ctx,
-        &shapes[0],
-        "soa://existing/subgroup/payments",
-        serde_json::json!({
-            "name": "Payments infrastructure",
-            "summary": "The team discussed dropped webhook retries during a recent payments outage."
-        }),
-    )
-    .await;
-
-        let ctx_map = existing_instance_context(&perspective, &shapes, None)
-            .await
-            .expect("existing_instance_context");
-        assert_eq!(ctx_map.len(), 1, "one seeded instance; got {ctx_map:#?}");
-        let inst = ctx_map
-            .get("soa://existing/subgroup/payments")
-            .expect("subgroup keyed by its id");
-        assert_eq!(inst.title, "Payments infrastructure");
-        assert_eq!(
-            inst.properties.get("summary").map(String::as_str),
-            Some("The team discussed dropped webhook retries during a recent payments outage."),
-            "summary scalar must be populated; got {:?}",
-            inst.properties
-        );
-        // Identity value must not be duplicated under `properties` (it is
-        // already carried by `title`); the type flag must never leak in either
-        // (setter-managed, not LLM-visible state).
-        assert!(!inst.properties.contains_key("name"));
-        assert!(!inst.properties.contains_key("type"));
     }
 }
